GitHub check in
Please let us know if this is not the proper way to submit our code changes.

We wanted to try writing code that would tell the rovers to move around the field in a certain pattern, instead of randomly. Now, with slight code changes, the rovers tend to move in a somewhat loose pattern clockwise around the field. The code changes are more of a proof of concept of these patterns than practical application.